cheking props  in externa; file
