+++ src/components/burgerconstructor/burgerconstructor.jsx
@@ -1,5 +1,6 @@
 import React, { useEffect } from "react";
 import PropTypes from "prop-types";
+import ingredientPropTypes from "../../utils/prop-types";
 import styles from "./burgerconstructor.module.css";
 import {
   Tab,
@@ -121,21 +122,6 @@ function BurgerConstructor(props) {
     </div>
   );
 }
-
-const ingredientPropTypes = PropTypes.shape({
-  calories: PropTypes.number.isRequired,
-  carbohydrates: PropTypes.number.isRequired,
-  fat: PropTypes.number.isRequired,
-  image: PropTypes.string.isRequired,
-  image_large: PropTypes.string.isRequired,
-  image_mobile: PropTypes.string.isRequired,
-  name: PropTypes.string.isRequired,
-  price: PropTypes.number.isRequired,
-  proteins: PropTypes.number.isRequired,
-  type: PropTypes.string.isRequired,
-  __v: PropTypes.number.isRequired,
-  _id: PropTypes.string.isRequired,
-});
 
 BurgerConstructor.propTypes = {
   ingredients: PropTypes.arrayOf(ingredientPropTypes),
