+++ src/components/burgeringredients/burgeringredients.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import styles from "./burgeringredients.module.css";
+import ingredientPropTypes from "../../utils/prop-types";
 import {
   ConstructorElement,
   CurrencyIcon,
@@ -73,21 +74,6 @@ function BurgerIngredients(props) {
     </>
   );
 }
-
-const ingredientPropTypes = PropTypes.shape({
-  calories: PropTypes.number.isRequired,
-  carbohydrates: PropTypes.number.isRequired,
-  fat: PropTypes.number.isRequired,
-  image: PropTypes.string.isRequired,
-  image_large: PropTypes.string.isRequired,
-  image_mobile: PropTypes.string.isRequired,
-  name: PropTypes.string.isRequired,
-  price: PropTypes.number.isRequired,
-  proteins: PropTypes.number.isRequired,
-  type: PropTypes.string.isRequired,
-  __v: PropTypes.number.isRequired,
-  _id: PropTypes.string.isRequired,
-});
 
 BurgerIngredients.propTypes = {
   ingredients: PropTypes.arrayOf(ingredientPropTypes),
