+++ src/components/ingredientdetails/ingredientdetails.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
+import ingredientPropTypes from "../../utils/prop-types";
 function IngredientDetails(props) {
   return (
     <section className="modal-ingredient-details">
@@ -40,21 +41,6 @@ function IngredientDetails(props) {
     </section>
   );
 }
-
-const ingredientPropTypes = PropTypes.shape({
-  calories: PropTypes.number.isRequired,
-  carbohydrates: PropTypes.number.isRequired,
-  fat: PropTypes.number.isRequired,
-  image: PropTypes.string.isRequired,
-  image_large: PropTypes.string.isRequired,
-  image_mobile: PropTypes.string.isRequired,
-  name: PropTypes.string.isRequired,
-  price: PropTypes.number.isRequired,
-  proteins: PropTypes.number.isRequired,
-  type: PropTypes.string.isRequired,
-  __v: PropTypes.number.isRequired,
-  _id: PropTypes.string.isRequired,
-});
 
 PropTypes.propTypes = {
   ingredientDetail: PropTypes.ingredientPropTypes,
