+++ src/utils/prop-types.jsx
@@ -1,0 +1,17 @@
+import PropTypes from "prop-types";
+const ingredientPropTypes = PropTypes.shape({
+  calories: PropTypes.number.isRequired,
+  carbohydrates: PropTypes.number.isRequired,
+  fat: PropTypes.number.isRequired,
+  image: PropTypes.string.isRequired,
+  image_large: PropTypes.string.isRequired,
+  image_mobile: PropTypes.string.isRequired,
+  name: PropTypes.string.isRequired,
+  price: PropTypes.number.isRequired,
+  proteins: PropTypes.number.isRequired,
+  type: PropTypes.string.isRequired,
+  __v: PropTypes.number.isRequired,
+  _id: PropTypes.string.isRequired,
+});
+
+export default ingredientPropTypes;
